requirements.txt
fastapi
uvicorn
newspaper3k
supabase
lxml_html_clean
google-generativeai
tavily-python
